backend/requirements.txt
fastapi[all]==0.115.11
sqlalchemy[asyncio]==2.0.39
databases[mysql]==0.9.0
pydantic_settings==2.8.1
alembic==1.15.1
gunicorn==23.0.0
uvicorn[standard]==0.34.0
asyncmy==0.2.10
redis==6.2.0
passlib==1.7.4
python-jose[cryptography]==3.4.0
bcrypt==4.3.0
motor==3.7.0
pydantic==2.10.6
aiokafka==0.12.0
websockets==15.0.1
aioboto3==13.3.0
pdf2image==1.17.0
pymilvus==2.5.6
pillow==11.1.0
openai==1.66.3
langchain-core>=0.3.0,<0.4.0
tenacity==9.0.0
sse_starlette==2.2.1
docker==7.1.0
mcp==1.9.1
